requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
django-phonenumber-field==7.1.0
phonenumbers==8.13.19
Pillow==10.0.0
django-ckeditor==6.7.0
